+++ th2_data_services/provider/v5/commands/grpc.py
@@ -429,7 +429,6 @@ class GetMessagesGRPCObject(IGRPCProvider5Command, ProviderAdaptableCommand):
         message_id: List[str] = None,
         attached_events: bool = False,
         filters: List[Filter] = None,
-        lookup_limit_days: int = None,
     ):
         """GetMessagesGRPCObject constructor.
 
@@ -445,7 +444,6 @@ class GetMessagesGRPCObject(IGRPCProvider5Command, ProviderAdaptableCommand):
             message_id: List of message ids to restore the search
             attached_events: If true, it will additionally load attachedEventsIds.
             filters: Filters using in search for messages.
-            lookup_limit_days: The number of days that will be viewed on the first request to get the one closest to the specified timestamp. By default it is absent - not limited to the past and up to the present moment to the future.
         """
         super().__init__()
         self._start_timestamp = start_timestamp
@@ -458,7 +456,6 @@ class GetMessagesGRPCObject(IGRPCProvider5Command, ProviderAdaptableCommand):
         self._filters = filters
         self._message_id = message_id
         self._attached_events = attached_events
-        self._lookup_limit_days = lookup_limit_days
 
     def handle(self, data_source: GRPCProvider5DataSource) -> List[MessageData]:
         api = data_source.source_api
@@ -477,7 +474,6 @@ class GetMessagesGRPCObject(IGRPCProvider5Command, ProviderAdaptableCommand):
             filters=self._filters,
             message_id=self._message_id,
             attached_events=self._attached_events,
-            lookup_limit_days=self._lookup_limit_days,
         )
         for response in stream_response:
             if response.WhichOneof("data") == "message":
@@ -507,7 +503,6 @@ class GetMessages(IGRPCProvider5Command, ProviderAdaptableCommand):
         message_id: List[str] = None,
         attached_events: bool = False,
         cache: bool = False,
-        lookup_limit_days: int = None,
     ):
         """GetMessages constructor.
 
@@ -524,7 +519,6 @@ class GetMessages(IGRPCProvider5Command, ProviderAdaptableCommand):
             message_id: List of message ids to restore the search
             attached_events: If true, it will additionally load attachedEventsIds.
             cache: If True, all requested data from rpt-data-provider will be saved to cache.
-            lookup_limit_days: The number of days that will be viewed on the first request.
         """
         super().__init__()
         self._start_timestamp = start_timestamp
@@ -538,7 +532,6 @@ class GetMessages(IGRPCProvider5Command, ProviderAdaptableCommand):
         self._message_id = message_id
         self._attached_events = attached_events
         self._cache = cache
-        self.lookup_limit_days = lookup_limit_days
 
         self._decoder = GRPCObjectToDictAdapter()
         self._wrapper_deleter = DeleteMessageWrappersAdapter()
@@ -559,7 +552,6 @@ class GetMessages(IGRPCProvider5Command, ProviderAdaptableCommand):
             message_id=self._message_id,
             attached_events=self._attached_events,
             filters=self._filters,
-            lookup_limit_days=self.lookup_limit_days,
         ).handle(data_source)
         for message in stream:
             message = self._decoder.handle(message)